chore(example): Reorganize ToolKit section and bump version to 5.2.2
- Move StepDialog from Controls section to ToolKit section
- Add NoticeDialog to ToolKit section navigation
- Reorganize ToolKit navigation items for better grouping
- Update package version from 5.2.1-fix to 5.2.2
- Update assembly and file versions to 5.2.2.0
- Align control page factory with updated navigation structure
