+++ samples/PleasantUI.Example/Pages/ToolKit/StepDialogPage.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using PleasantUI.Example.Views.Pages.PleasantControlPages;
 
-namespace PleasantUI.Example.Pages.PleasantControls;
+namespace PleasantUI.Example.Pages.Toolkit;
 
 public class StepDialogPage : LocalizedPage
 {
